legacy/old_models/tf2/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
absl-py==1.4.0
    # via
    #   array-record
    #   tensorboard
    #   tensorflow
    #   tensorflow-datasets
    #   tensorflow-metadata
    #   tensorflow-model-optimization
    #   tf-slim
array-record==0.5.1
    # via tensorflow-datasets
astunparse==1.6.3
    # via tensorflow
beartype==0.17.2
    # via picsellia
bleach==6.2.0
    # via kaggle
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   kaggle
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via tensorflow-datasets
colorama==0.4.6
    # via sacrebleu
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via
    #   lvis
    #   matplotlib
cython==3.0.11
    # via
    #   lvis
    #   tf-models-official
deprecation==2.1.0
    # via picsellia
dm-tree==0.1.8
    # via
    #   tensorflow-datasets
    #   tensorflow-model-optimization
etils==1.11.0
    # via
    #   array-record
    #   tensorflow-datasets
flatbuffers==24.3.25
    # via tensorflow
fonttools==4.55.0
    # via matplotlib
fsspec==2024.10.0
    # via etils
gast==0.4.0
    # via tensorflow
gin-config==0.5.0
    # via tf-models-official
google-api-core==2.23.0
    # via google-api-python-client
google-api-python-client==2.154.0
    # via tf-models-official
google-auth==2.36.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   tensorboard
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.0.0
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
googleapis-common-protos==1.63.1
    # via
    #   google-api-core
    #   tensorflow-metadata
grpcio==1.68.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.12.1
    # via tensorflow
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==3.10
    # via requests
immutabledict==4.2.1
    # via tf-models-official
importlib-resources==6.4.5
    # via etils
joblib==1.4.2
    # via scikit-learn
kaggle==1.6.17
    # via tf-models-official
keras==2.13.1
    # via tensorflow
kiwisolver==1.4.7
    # via
    #   lvis
    #   matplotlib
libclang==18.1.1
    # via tensorflow
lvis==0.5.3
    # via -r requirements.in
lxml==5.3.0
    # via sacrebleu
markdown==3.7
    # via tensorboard
markupsafe==3.0.2
    # via werkzeug
matplotlib==3.9.2
    # via
    #   lvis
    #   pycocotools
    #   tf-models-official
numpy==1.23.5
    # via
    #   contourpy
    #   etils
    #   h5py
    #   lvis
    #   matplotlib
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   pycocotools
    #   sacrebleu
    #   scikit-learn
    #   scipy
    #   seqeval
    #   tensorboard
    #   tensorflow
    #   tensorflow-datasets
    #   tensorflow-hub
    #   tensorflow-model-optimization
    #   tf-models-official
oauth2client==4.1.3
    # via tf-models-official
oauthlib==3.2.2
    # via requests-oauthlib
opencv-python==4.10.0.84
    # via lvis
opencv-python-headless==4.10.0.84
    # via tf-models-official
opt-einsum==3.4.0
    # via tensorflow
orjson==3.10.12
    # via picsellia
packaging==24.2
    # via
    #   deprecation
    #   matplotlib
    #   tensorflow
pandas==2.2.3
    # via tf-models-official
picsellia==6.15.0
    # via -r requirements.in
picsellia-annotations==0.6.0
    # via picsellia
picsellia-connexion-services==0.2.1
    # via picsellia
pillow==9.5.0
    # via
    #   matplotlib
    #   picsellia
    #   tf-models-official
portalocker==3.0.0
    # via sacrebleu
promise==2.3
    # via tensorflow-datasets
proto-plus==1.25.0
    # via google-api-core
protobuf==4.25.5
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   tensorboard
    #   tensorflow
    #   tensorflow-datasets
    #   tensorflow-hub
    #   tensorflow-metadata
psutil==6.1.0
    # via
    #   tensorflow-datasets
    #   tf-models-official
py-cpuinfo==9.0.0
    # via tf-models-official
pyasn1==0.6.1
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via
    #   google-auth
    #   oauth2client
pycocotools==2.0.8
    # via tf-models-official
pydantic==1.10.19
    # via
    #   picsellia
    #   picsellia-annotations
pyparsing==3.2.0
    # via
    #   httplib2
    #   lvis
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   kaggle
    #   lvis
    #   matplotlib
    #   pandas
python-slugify==8.0.4
    # via kaggle
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   picsellia
    #   tf-models-official
regex==2024.11.6
    # via sacrebleu
requests==2.32.3
    # via
    #   google-api-core
    #   kaggle
    #   picsellia
    #   picsellia-connexion-services
    #   requests-oauthlib
    #   tensorboard
    #   tensorflow-datasets
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
sacrebleu==2.4.3
    # via tf-models-official
scikit-learn==1.5.2
    # via seqeval
scipy==1.14.1
    # via
    #   -r requirements.in
    #   scikit-learn
    #   tf-models-official
sentencepiece==0.2.0
    # via tf-models-official
seqeval==1.2.2
    # via tf-models-official
setuptools==75.6.0
    # via
    #   tensorboard
    #   tensorflow
six==1.16.0
    # via
    #   astunparse
    #   google-pasta
    #   kaggle
    #   lvis
    #   oauth2client
    #   promise
    #   python-dateutil
    #   tensorflow
    #   tensorflow-model-optimization
    #   tf-models-official
tabulate==0.9.0
    # via sacrebleu
tdqm==0.0.1
    # via picsellia
tensorboard==2.13.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.13.0
    # via
    #   -r requirements.in
    #   tensorflow-text
    #   tf-models-official
tensorflow-datasets==4.9.0
    # via tf-models-official
tensorflow-estimator==2.13.0
    # via tensorflow
tensorflow-hub==0.16.1
    # via
    #   tensorflow-text
    #   tf-models-official
tensorflow-io==0.37.1
    # via -r requirements.in
tensorflow-io-gcs-filesystem==0.37.1
    # via
    #   tensorflow
    #   tensorflow-io
tensorflow-metadata==1.13.1
    # via tensorflow-datasets
tensorflow-model-optimization==0.8.0
    # via tf-models-official
tensorflow-text==2.13.0
    # via tf-models-official
termcolor==2.5.0
    # via
    #   tensorflow
    #   tensorflow-datasets
text-unidecode==1.3
    # via python-slugify
tf-keras==2.15.0
    # via tensorflow-hub
tf-models-official==2.13.2
    # via -r requirements.in
tf-slim==1.1.0
    # via
    #   -r requirements.in
    #   tf-models-official
threadpoolctl==3.5.0
    # via scikit-learn
toml==0.10.2
    # via tensorflow-datasets
tqdm==4.67.1
    # via
    #   kaggle
    #   tdqm
    #   tensorflow-datasets
typing-extensions==4.5.0
    # via
    #   etils
    #   pydantic
    #   tensorflow
tzdata==2024.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via
    #   kaggle
    #   requests
webencodings==0.5.1
    # via bleach
werkzeug==3.1.3
    # via tensorboard
wheel==0.45.1
    # via
    #   astunparse
    #   tensorboard
wrapt==1.14.1
    # via
    #   tensorflow
    #   tensorflow-datasets
xmltodict==0.13.0
    # via picsellia-annotations
zipp==3.21.0
    # via etils
